abstract dup adapter mocks to dedicated mock package
