--- examples/load_pub_sources.go
@@ -1,1 +1,0 @@
-package examples
